+++ lib/misskey_auth.dart
@@ -1,8 +1,10 @@
 // Models
 export 'src/models/oauth_models.dart';
+export 'src/models/miauth_models.dart';
 
 // API
 export 'src/api/misskey_oauth_client.dart';
+export 'src/api/misskey_miauth_client.dart';
 
 // Exceptions
 export 'src/exceptions/misskey_auth_exception.dart';